requirements.txt
axme>=0.1.2
pyautogen>=0.4.0
openai>=1.60.0
